solved issue with plotting one point
The problem was that I was doing self.current_noncentral_data[key] = self.all_noncentral_data[key][timestep]. What this was doing was setting the array (since  self.current_noncentral_data[key] is an array) to only a slice of the array. This array is shared between the two attributes, so by doing that I was modifying the array in place, causing issues later. The proper handling for that is to empty the self.current_noncentral_data dictionary ( set to {}) and then fill it in with the slice of the array we want.