working version, although slight but with the checkboxes
For some reason sometimes the default color checkbox as well as the individual point checkbox stay on even when the cmap checkbox is on. However, they should be off because I am setting them to be unckecked when the cmap checkbox is checked. This does not affect the picture though